requirements.txt
django
django-cors-headers
django-allauth
djangorestframework
dj-rest-auth
djangorestframework-simplejwt
django-filter